+++ bin/fcgi-bin/nph-genericinterface.pl
@@ -37,7 +37,11 @@ use Kernel::System::ObjectManager;
 # response loop
 while ( my $WebRequest = new CGI::Fast ) {
 
-    local $Kernel::OM = Kernel::System::ObjectManager->new();
+    local $Kernel::OM = Kernel::System::ObjectManager->new(
+        'Kernel::System::Web::Request' => {
+            WebRequest => $WebRequest,
+            }
+    );
     my $Provider = Kernel::GenericInterface::Provider->new();
 
     $Provider->Run();